Remove Puppet Gem Dependency
`librarian-puppet` has a hard dependency on the Puppet _gem_ in the
gemspec and requires it in `lib/librarian/puppet.rb`. In fact, there
is no hard dependency in the code itself that requires the Puppet source
to be available, and many installations do not use the gem (specifically
Windows).

`librarian-puppet` only requires Puppet because it shells out to it in
certain cases.

On Windows, Mac, and most Linux, there are binary packages
available, and I prefer these. I also like to use librarian-puppet. On
Windows, I can't do this because the gem doesn't work in cygwin, and I
currently have to run librarian-puppet in Cygwin on Windows.

In the future I recommend adding a check for the "puppet" binary on the
PATH if you need to shell out, but removing this dep doesn't break
anything otherwise.